Crate fonts_traits

Modules

font_descriptor 🔒

font_identifier 🔒

font_template 🔒

system_font_service_proxy 🔒

Structs

CSSFontFaceDescriptors

This data structure represents the various optional descriptors that can be applied to a @font-face rule in CSS. These are used to create a [FontTemplate] from the given font data used as the source of the @font-face rule. If values like weight, stretch, and style are not specified they are initialized based on the contents of the font itself.

FontData

A data structure to store data for fonts. Data is stored internally in an GenericSharedMemory handle, so that it can be sent without serialization across IPC channels.

FontDataAndIndex

Raw font data and an index

FontDescriptor

FontDescriptor describes the parameters of a Font. It represents rendering a given font template at a particular size, with a particular font-variant-caps applied, etc. This contrasts with FontTemplateDescriptor in that the latter represents only the parameters inherent in the font data (weight, stretch, etc.).

FontFaceRuleInfo

FontTemplate

This describes all the information needed to create font instance handles. It contains a unique FontTemplateData structure that is platform specific.

FontTemplateDescriptor

Describes how to select a font from a given family. This is very basic at the moment and needs to be expanded or refactored when we support more of the font styling parameters.

FontTemplateRef

A reference to a FontTemplate with shared ownership and mutability.

LocalFontIdentifier

An identifier for a local font on systems using Freetype.

LowercaseFontFamilyName

SystemFontServiceProxy

The public interface to the [SystemFontService], used by per-Document FontContext instances.

SystemFontServiceProxySender

WebFontSetDifference

Describes how the set of active @font-face rules was changed after a call to FontContext::rebuild_font_face_set.

Enums

FontDataError

FontIdentifier

SystemFontServiceMessage

Commands that the FontContext sends to the SystemFontService.

WebFontLoadEvent

Traits

FontTemplateRefMethods

IsOblique

Type Aliases

StylesheetWebFontLoadFinishedCallback