fonts_traits

Struct FontDescriptor

pub struct FontDescriptor {
    pub weight: FontWeight,
    pub stretch: FontStretch,
    pub style: FontStyle,
    pub variant: T,
    pub pt_size: Au,
    pub variation_settings: Vec<FontVariation>,
    pub synthesis_weight: FontSynthesis,
    pub optical_sizing: T,
}

FontDescriptor describes the parameters of a Font. It represents rendering a given font template at a particular size, with a particular font-variant-caps applied, etc. This contrasts with FontTemplateDescriptor in that the latter represents only the parameters inherent in the font data (weight, stretch, etc.).

Fields
weight: FontWeightstretch: FontStretchstyle: FontStylevariant: Tpt_size: Auvariation_settings: Vec<FontVariation>

The value of the @font-variation-settings property.

This does not include synthesized variations from font-style, font-stretch etc.

synthesis_weight: FontSynthesisoptical_sizing: T